# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Email marketplace platform API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: number;
  email: string;
  walletBalance: number;
}

export interface UserProfile {
  id: number;
  email: string;
  walletBalance: number;
  totalSubmissions: number;
  approvedSubmissions: number;
  pendingSubmissions: number;
}

export interface UserInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Submission {
  id: number;
  email: string;
  status: SubmissionStatus;
  pricePaid: number;
  createdAt: string;
}

export interface SubmissionInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type AdminSubmissionStatus = typeof AdminSubmissionStatus[keyof typeof AdminSubmissionStatus];


export const AdminSubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface AdminSubmission {
  id: number;
  userId: number;
  userEmail: string;
  email: string;
  password: string;
  status: AdminSubmissionStatus;
  pricePaid: number;
  createdAt: string;
}

export type SubmissionStatusUpdateStatus = typeof SubmissionStatusUpdateStatus[keyof typeof SubmissionStatusUpdateStatus];


export const SubmissionStatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SubmissionStatusUpdate {
  status: SubmissionStatusUpdateStatus;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface Withdrawal {
  id: number;
  amount: number;
  telebirrNumber: string;
  telebirrName: string;
  status: WithdrawalStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export interface WithdrawalInput {
  /** @minimum 1 */
  amount: number;
  /** @minLength 10 */
  telebirrNumber: string;
  /** @minLength 1 */
  telebirrName: string;
}

export type AdminWithdrawalStatus = typeof AdminWithdrawalStatus[keyof typeof AdminWithdrawalStatus];


export const AdminWithdrawalStatus = {
  pending: 'pending',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface AdminWithdrawal {
  id: number;
  userId: number;
  userEmail: string;
  amount: number;
  telebirrNumber: string;
  telebirrName: string;
  status: AdminWithdrawalStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export type WithdrawalStatusUpdateStatus = typeof WithdrawalStatusUpdateStatus[keyof typeof WithdrawalStatusUpdateStatus];


export const WithdrawalStatusUpdateStatus = {
  completed: 'completed',
  rejected: 'rejected',
} as const;

export interface WithdrawalStatusUpdate {
  status: WithdrawalStatusUpdateStatus;
  adminNote?: string;
}

export interface AdminUser {
  id: number;
  email: string;
  walletBalance: number;
  totalSubmissions: number;
  approvedSubmissions: number;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalSubmissions: number;
  pendingSubmissions: number;
  approvedSubmissions: number;
  totalEmailsBought: number;
  totalPayoutsBirr: number;
  pendingWithdrawals: number;
}

export interface PlatformSettings {
  pricePerEmail: number;
}

export interface SettingsUpdate {
  /** @minimum 1 */
  pricePerEmail: number;
}